+++ src/code_indexer/server/services/claude_cli_manager.py
@@ -24,6 +24,7 @@ from typing import Callable, Optional, List, Tuple
 
 from code_indexer.utils.file_locking import nfs_safe_flock, nfs_safe_funlock
 from code_indexer.server.logging_utils import format_error_log
+from code_indexer.utils.subprocess_env import build_cidx_subprocess_env
 
 logger = logging.getLogger(__name__)
 
@@ -562,6 +563,7 @@ class ClaudeCliManager:
                 ["cidx", "index"],
                 cwd=str(self._meta_dir),
                 capture_output=True,
+                env=build_cidx_subprocess_env(),
                 check=False,
             )
             logger.info(